refactor: better folder structure for components
