+++ components/blog/BlogPostSummary.tsx
@@ -1,5 +1,5 @@
 //? Require the interface to ensure we receive the proper data
-import BlogPostSummaryProperties from "../types/BlogPostSummaryProperties.ts";
+import BlogPostSummaryProperties from "../../types/BlogPostSummaryProperties.ts";
 
 //? Exports a single Blog Post Summary
 export function BlogPostSummary(summary: BlogPostSummaryProperties) {
