+++ components/home/HomeContent.tsx
