+++ routes/[blog]/[post].tsx
@@ -2,11 +2,11 @@
 // importing this file from /services/fetchPosts.ts
 
 // //? Head component with all Meta tags pre-set
-// import { CustomHead } from "../../components/CustomHead.tsx";
+// import { CustomHead } from "../../components/base/CustomHead.tsx";
 // //? Lateral text with theme switcher
-// import { Base } from "../../components/Base.tsx";
+// import { Base } from "../../components/base/Base.tsx";
 // //? Navigation Buttons to go back to the previous page or to the next page (optional)
-// import BlogNavigationButtons from "../../components/BlogNavigationButtons.tsx";
+// import BlogNavigationButtons from "../../components/blog/BlogNavigationButtons.tsx";
 // //? To know what is the current route
 // import { Handlers } from "$fresh/server.ts";
 // //? Parse content from a file into JSX
