+++ routes/[blog]/how-create-theme-switcher-deno-fresh.tsx
@@ -1,11 +1,11 @@
 //? Head component with all Meta tags pre-set
-import { CustomHead } from "../../components/CustomHead.tsx";
+import { CustomHead } from "../../components/base/CustomHead.tsx";
 //? Lateral text with theme switcher
-import { Base } from "../../components/Base.tsx";
+import { Base } from "../../components/base/Base.tsx";
 //? Navigation Buttons to go back to the previous page or to the next page (optional)
-import BlogNavigationButtons from "../../components/BlogNavigationButtons.tsx";
+import BlogNavigationButtons from "../../components/blog/BlogNavigationButtons.tsx";
 //? A HTML Link component to pre-format links and reduce boiletplate
-import { GradientLink } from "../../components/GradientLink.tsx";
+import { GradientLink } from "../../components/base/GradientLink.tsx";
 
 export default function Home() {
   return (
