+++ routes/[toys]/insanity.tsx
@@ -1,9 +1,9 @@
 //? Create blog content inside Base component
-import { Base } from "../../components/Base.tsx";
+import { Base } from "../../components/base/Base.tsx";
 //? Head component with all Meta tags pre-set
-import { CustomHead } from "../../components/CustomHead.tsx";
+import { CustomHead } from "../../components/base/CustomHead.tsx";
 //? Navigation Buttons to go back to the previous page or to the next article
-import BlogNavigationButtons from "../../components/BlogNavigationButtons.tsx";
+import BlogNavigationButtons from "../../components/blog/BlogNavigationButtons.tsx";
 //? Infinitely expandable insanity section
 import InsanitySection from "../../islands/InsanitySection.tsx";
