+++ routes/me.tsx
@@ -1,11 +1,11 @@
 //? Create blog content inside Base component
-import { Base } from "../components/Base.tsx";
+import { Base } from "../components/base/Base.tsx";
 //? Head component with all Meta tags pre-set
-import { CustomHead } from "../components/CustomHead.tsx";
+import { CustomHead } from "../components/base/CustomHead.tsx";
 //? A HTML Link component to pre-format links and reduce boiletplate
-import { GradientLink } from "../components/GradientLink.tsx";
+import { GradientLink } from "../components/base/GradientLink.tsx";
 //? Navigation Buttons to go back to the previous page or to the next article
-import BlogNavigationButtons from "../components/BlogNavigationButtons.tsx";
+import BlogNavigationButtons from "../components/blog/BlogNavigationButtons.tsx";
 
 export default function Home() {
   return (
